All message responses are now embeds

+++ src/commands/boosters/say.ts
@@ -14,6 +14,9 @@ export const command: Command = {
     // eslint-disable-next-line sort-keys
     run: async (client, msg, args) => {
 
+        args.shift();
+        if (args.length === 0) return client.embedReply(msg, { embed: { description: "No message specified" } });
+
         await client.reply(msg, { content: args.join(" "), "files": [...msg.attachments.values()] });
         await msg.delete();
     }
